+++ test/org/ojalgo/array/LongToNumberMapTest.java
@@ -21,14 +21,12 @@
  */
 package org.ojalgo.array;
 
+import org.junit.Test;
 import java.util.Map.Entry;
 import java.util.Random;
 import java.util.SortedMap;
 import java.util.TreeMap;
 
-import org.junit.Test;
-import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.Tags;
 import org.ojalgo.TestUtils;
 import org.ojalgo.constant.PrimitiveMath;
 import org.ojalgo.type.context.NumberContext;
@@ -38,19 +36,21 @@ import org.ojalgo.type.context.NumberContext;
  *
  * @author apete
  */
-public class LongToNumberMapTest  {
+public class LongToNumberMapTest {
 
     private static final NumberContext CONTEXT = new NumberContext();
     private static final Random RANDOM = new Random();
 
-    @Test public void testAlignCapacity() {
+    @Test
+    public void testAlignCapacity() {
         TestUtils.assertEquals(1, 1L << PrimitiveMath.powerOf2Larger(-1L));
         TestUtils.assertEquals(16, 1L << PrimitiveMath.powerOf2Larger(16L));
         TestUtils.assertEquals(512, 1L << PrimitiveMath.powerOf2Larger(365L));
         TestUtils.assertEquals(16_384, 1L << PrimitiveMath.powerOf2Larger(16_384L));
     }
 
-    @Test public void testCompareWithTreeMap() {
+    @Test
+    public void testCompareWithTreeMap() {
 
         final LongToNumberMap<Double> primit64Map = LongToNumberMap.factory(Primitive64Array.FACTORY).make();
         final LongToNumberMap<Double> direct64Map = LongToNumberMap.factory(BufferArray.DIRECT64).make();
@@ -68,7 +68,7 @@ public class LongToNumberMapTest  {
 
         for (int c = 0; c < 1_000; c++) {
 
-            final Long index = (long) RANDOM.nextInt(1_000);
+            final Long index = Long.valueOf(RANDOM.nextInt(1_000));
             final Double value = RANDOM.nextDouble();
 
             primit64Map.put(index, value);
@@ -78,7 +78,7 @@ public class LongToNumberMapTest  {
 
         for (int c = 0; c < 1_000; c++) {
 
-            final Long index = (long) RANDOM.nextInt(1_000);
+            final Long index = Long.valueOf(RANDOM.nextInt(1_000));
 
             primit64Map.remove(index);
             direct64Map.remove(index);
@@ -95,7 +95,7 @@ public class LongToNumberMapTest  {
         TestUtils.assertEquals(expectedMap.size(), direct64Map.size());
 
         for (final Entry<Long, Double> entry : expectedMap.entrySet()) {
-            final double expectedValue = entry.getValue();
+            final double expectedValue = entry.getValue().doubleValue();
             TestUtils.assertEquals(expectedValue, primit64Map.get(entry.getKey()).doubleValue(), CONTEXT);
             TestUtils.assertEquals(expectedValue, primit64Map.doubleValue(entry.getKey()), CONTEXT);
             TestUtils.assertEquals(expectedValue, direct64Map.get(entry.getKey()).doubleValue(), CONTEXT);
@@ -104,7 +104,8 @@ public class LongToNumberMapTest  {
 
     }
 
-    @Test public void testSubmap() {
+    @Test
+    public void testSubmap() {
 
         final LongToNumberMap<Double> tmpMap = LongToNumberMap.factory(Primitive64Array.FACTORY).make();
